requirements.txt
fastapi==0.131.0
uvicorn==0.41.0
sqlalchemy==2.0.46
python-jose==3.5.0
email-validator==2.3.0
pydantic_settings==2.13.1
python-multipart==0.0.22
pytest==9.0.2
bcrypt==4.0.1
passlib==1.7.4